All Batch Rendering Stuff Done. Need to do some general clean up. Very Happy!

+++ BulletHell-Engine/src/include/Attack.hpp
@@ -5,7 +5,7 @@
 
 class Game;
 struct AttackPattern {
-	uint32_t bulletDataIndex;
+	uint16_t bulletDataIndex;
 	FireTypes fireType;
 	sf::Clock timeSinceShot;
 	int bulletsFired;
